[#3933] Refactor ApplicationContext  - Refactor PluginLoader
