+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/instrument/classloading/PlainClassLoaderHandler.java
@@ -129,7 +129,7 @@ public class PlainClassLoaderHandler implements ClassInjector {
 
 
 
-    private Class<?> injectClass0(ClassLoader classLoader, String className) throws IllegalArgumentException, IllegalAccessException, InvocationTargetException {
+    private Class<?> injectClass0(ClassLoader classLoader, String className) throws IllegalArgumentException {
         if (isDebug) {
             logger.debug("Inject class className:{} cl:{}", className, classLoader);
         }
@@ -147,7 +147,7 @@ public class PlainClassLoaderHandler implements ClassInjector {
 
     }
 
-    private InputStream getInputStream(ClassLoader classLoader, String classPath) throws IllegalArgumentException, IllegalAccessException, InvocationTargetException {
+    private InputStream getInputStream(ClassLoader classLoader, String classPath) throws IllegalArgumentException {
         if (isDebug) {
             logger.debug("Get input stream className:{} cl:{}", classPath, classLoader);
