+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/plugin/ClassNameFilterChain.java
@@ -24,13 +24,13 @@ import java.util.List;
  */
 public class ClassNameFilterChain implements ClassNameFilter {
 
-    private final List<ClassNameFilter> filterChain;
+    private final ClassNameFilter[] filterChain;
 
     public ClassNameFilterChain(List<ClassNameFilter> filterChain) {
         if (filterChain == null) {
             throw new NullPointerException("filterChain must not be null");
         }
-        this.filterChain = new ArrayList<ClassNameFilter>(filterChain);
+        this.filterChain = filterChain.toArray(new ClassNameFilter[filterChain.size()]);
     }
