+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/plugin/PluginConfig.java
@@ -16,13 +16,11 @@
 
 package com.navercorp.pinpoint.profiler.plugin;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.navercorp.pinpoint.common.plugin.JarPlugin;
+import com.navercorp.pinpoint.common.plugin.Plugin;
+import com.navercorp.pinpoint.common.util.Assert;
 
-import java.io.File;
-import java.io.IOException;
-import java.net.URI;
-import java.net.URISyntaxException;
+
 import java.net.URL;
 import java.util.Collections;
 import java.util.List;
@@ -33,54 +31,47 @@ import java.util.jar.JarFile;
  */
 public class PluginConfig {
 
-    private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
     public static final String PINPOINT_PLUGIN_PACKAGE = "Pinpoint-Plugin-Package";
     public static final List<String> DEFAULT_PINPOINT_PLUGIN_PACKAGE_NAME = Collections.singletonList("com.navercorp.pinpoint.plugin");
 
-    private final URL pluginJar;
-    private final JarFile pluginJarFile;
+    private final Plugin plugin;
+    private final JarFile pluginJar;
+
     private String pluginJarURLExternalForm;
 
     private final ClassNameFilter pluginPackageFilter;
 
-    public PluginConfig(URL pluginJar, ClassNameFilter pluginPackageFilter) {
-        if (pluginJar == null) {
-            throw new NullPointerException("pluginJar must not be null");
-        }
-        this.pluginJar = pluginJar;
-        this.pluginJarFile = createJarFile(pluginJar);
+    public PluginConfig(Plugin plugin, ClassNameFilter pluginPackageFilter) {
+        this.plugin = Assert.requireNonNull(plugin, "plugin must not be null");
 
         this.pluginPackageFilter = pluginPackageFilter;
+        this.pluginJar = getJarFile(plugin);
+
+    }
+
+    private JarFile getJarFile(Plugin plugin) {
+        if (plugin instanceof JarPlugin) {
+            return ((JarPlugin) plugin).getJarFile();
+        }
+        throw new IllegalArgumentException("unsupported plugin " + plugin);
     }
 
 
     public URL getPluginJar() {
-        return pluginJar;
+        return plugin.getURL();
     }
 
     public JarFile getPluginJarFile() {
-        return pluginJarFile;
+        return pluginJar;
     }
 
     public String getPluginJarURLExternalForm() {
         if (this.pluginJarURLExternalForm == null) {
-            this.pluginJarURLExternalForm = pluginJar.toExternalForm();
+            this.pluginJarURLExternalForm = plugin.getURL().toExternalForm();
         }
         return this.pluginJarURLExternalForm;
     }
-
-    private JarFile createJarFile(URL pluginJar) {
-        try {
-            final URI uri = pluginJar.toURI();
-            return new JarFile(new File(uri));
-        } catch (URISyntaxException e) {
-            throw new RuntimeException("URISyntax error. " + e.getCause(), e);
-        } catch (IOException e) {
-            throw new RuntimeException("IO error. " + e.getCause(), e);
-        }
-    }
-
 
     public ClassNameFilter getPluginPackageFilter() {
         return pluginPackageFilter;
@@ -89,8 +80,7 @@ public class PluginConfig {
     @Override
     public String toString() {
         return "PluginConfig{" +
-                "pluginJar=" + pluginJar +
-                ", pluginJarFile=" + pluginJarFile +
+                "pluginJar=" + plugin.getURL() +
                 ", pluginJarURLExternalForm='" + pluginJarURLExternalForm + '\'' +
                 ", pluginPackageFilter=" + pluginPackageFilter +
                 '}';
