+++ profiler/src/test/java/com/navercorp/pinpoint/profiler/context/graph/DependencyGraph.java
@@ -24,15 +24,15 @@ import com.google.inject.grapher.graphviz.GraphvizModule;
 import com.navercorp.pinpoint.bootstrap.AgentOption;
 import com.navercorp.pinpoint.bootstrap.DefaultAgentOption;
 import com.navercorp.pinpoint.bootstrap.config.DefaultProfilerConfig;
+import com.navercorp.pinpoint.bootstrap.config.ProfilerConfig;
 import com.navercorp.pinpoint.common.Charsets;
-import com.navercorp.pinpoint.common.service.DefaultAnnotationKeyRegistryService;
-import com.navercorp.pinpoint.common.service.DefaultServiceTypeRegistryService;
 import com.navercorp.pinpoint.profiler.context.module.DefaultApplicationContext;
 import com.navercorp.pinpoint.profiler.context.module.InterceptorRegistryModule;
 import com.navercorp.pinpoint.profiler.context.module.ModuleFactory;
 import com.navercorp.pinpoint.profiler.context.module.OverrideModuleFactory;
 import com.navercorp.pinpoint.profiler.interceptor.registry.InterceptorRegistryBinder;
 import com.navercorp.pinpoint.profiler.util.TestInterceptorRegistryBinder;
+import org.mockito.Mockito;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -42,8 +42,10 @@ import java.io.PrintWriter;
 import java.lang.instrument.Instrumentation;
 import java.net.URL;
 import java.security.CodeSource;
+import java.util.Collections;
 
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.spy;
 
 /**
  * @author Woonduk Kang(emeroad)
@@ -77,12 +79,12 @@ public class DependencyGraph {
     }
 
     private DefaultApplicationContext newApplicationContext() {
-        DefaultProfilerConfig profilerConfig = new DefaultProfilerConfig();
-        profilerConfig.setStaticResourceCleanup(true);
+        ProfilerConfig profilerConfig = spy(new DefaultProfilerConfig());
+        Mockito.when(profilerConfig.getStaticResourceCleanup()).thenReturn(true);
 
         Instrumentation instrumentation = mock(Instrumentation.class);
         AgentOption agentOption = new DefaultAgentOption(instrumentation,
-                "mockAgent", "mockApplicationName", profilerConfig, new URL[0],
+                "mockAgent", "mockApplicationName", profilerConfig, Collections.<String>emptyList(),
                 null);
 
         InterceptorRegistryBinder interceptorRegistryBinder = new TestInterceptorRegistryBinder();
