+++ web/src/test/java/com/navercorp/pinpoint/web/vo/callstacks/RecordFactoryTest.java
@@ -33,7 +33,6 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import java.util.Collections;
-import java.util.List;
 
 /**
  * @author Woonduk Kang(emeroad)
